requirements.txt
requests
python-dotenv
schedule
snowflake-connector-python

